add comunication between ProductWebApi and OrderWebAPi

+++ OrderWebApi/Models/OrderDetail.cs
@@ -9,8 +9,8 @@ namespace OrderWebApi.Models
         [BsonElement("product_id"), BsonRepresentation(BsonType.Int32)]
         public int ProductId { get; set; }
 
-        [BsonElement("quantity"), BsonRepresentation(BsonType.Decimal128)]
-        public decimal Quantity { get; set; }
+        [BsonElement("quantity"), BsonRepresentation(BsonType.Int32)]
+        public int Quantity { get; set; }
 
         [BsonElement("unit_price"), BsonRepresentation(BsonType.Decimal128)]
         public decimal UnitPrice { get; set; }
